requirements.txt
fastapi
uvicorn
python-jose[cryptography]
PyJWT
python-multipart
python-dotenv
passlib[bcrypt]
alembic
sqlalchemy>=2.0
psycopg2-binary





